+++ database/settings/table.go
@@ -14,13 +14,13 @@ const (
 CREATE TABLE
 IF NOT EXISTS
 settings (
-	id				SERIAL PRIMARY KEY,
+	id				    BIGSERIAL PRIMARY KEY,
 	compiler			JSON DEFAULT NULL,
 	queue				JSON DEFAULT NULL,
-	repo_allowlist			VARCHAR(1000),
-	schedule_allowlist		VARCHAR(1000),
-	created_at			INTEGER,
-	updated_at			INTEGER,
+	repo_allowlist	    VARCHAR(1000),
+	schedule_allowlist	VARCHAR(1000),
+	created_at			BIGINT,
+	updated_at			BIGINT,
 	updated_by			VARCHAR(250)
 );
 `
